+++ app/src/main/java/com/example/sleepeasy/navigation/AppNavigationGraph.kt
@@ -4,8 +4,8 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.example.sleepeasy.screens.SleepFormScreen
-import com.example.sleepeasy.screens.SleepHubScreen
+import com.example.sleepeasy.ui.SleepFormScreen
+import com.example.sleepeasy.ui.SleepHubScreen
 
 @Composable
 fun AppNavigationGraph() {
